+++ src/matching_engine/event/fill_event.cpp
@@ -1,6 +1,6 @@
 #include "fill_event.h"
 
-FillEvent::FillEvent(uint64_t clientId, uint64_t orderId, uint64_t matchId, double price, double qty) : Event(clientId, orderId), _matchId(matchId), _price(price), _qty(qty) {
+FillEvent::FillEvent(uint64_t clientId, uint64_t orderId, uint64_t matchId, int64_t price, double qty) : Event(clientId, orderId), _matchId(matchId), _price(price), _qty(qty) {
     if (price <= 0) {
         throw std::invalid_argument("Price must be greater than 0");
     }
@@ -12,5 +12,5 @@ FillEvent::FillEvent(uint64_t clientId, uint64_t orderId, uint64_t matchId, doub
 EventType FillEvent::getType() const { return EventType::Fill; }
 
 uint64_t FillEvent::getMatchId() const { return _matchId; }
-double FillEvent::getPrice() const { return _price; }
+int64_t FillEvent::getPrice() const { return _price; }
 double FillEvent::getQty() const { return _qty; }
