+++ src/matching_engine/order/order.cpp
@@ -1,6 +1,6 @@
 #include "order.h"
 
-Order::Order(uint64_t clientId, uint64_t orderId, std::string symbol, Side side, double price, double qtyRemaining) : _clientId(clientId), _orderId(orderId), _symbol(symbol), _side(side), _price(price), _qtyRemaining(qtyRemaining) {
+Order::Order(uint64_t clientId, uint64_t orderId, std::string symbol, Side side, int64_t price, double qtyRemaining) : _clientId(clientId), _orderId(orderId), _symbol(symbol), _side(side), _price(price), _qtyRemaining(qtyRemaining) {
     if (price <= 0) {
         throw std::invalid_argument("Price must be greater than 0");
     }
@@ -16,7 +16,7 @@ uint64_t Order::getClientId() const { return _clientId; }
 uint64_t Order::getOrderId() const { return _orderId; }
 std::string Order::getSymbol() const { return _symbol; }
 Side Order::getSide() const { return _side; }
-double Order::getPrice() const { return _price; }
+int64_t Order::getPrice() const { return _price; }
 double Order::getQtyRemaining() const { return _qtyRemaining; }
 bool Order::isFilled() const { return _isFilled; }
